Begin separating filter from analysis.
Split out the code that determines whether an event is in the
window from the code that analyses events.  This allows running
the first part as a filter and avoiding the, sometimes
expensive, reconstruction that needs to come before the
analysis.

Still a mess.  Need to remove the timing logic from the
analysis module and decide whose responsibility it is
to complain about unsorted events and so forth.