Continued work separating responsibilities.
Updated all the fcls to use the two separate modules
for filtering and analysis.

Moved code around between the two modules to make them
correctly separated.  Both of them need to know about the
window time and length because the filter filters on it
and the analysis module places events into histograms
based on it.