Rename NDMeV to MeV.
Because it isn't really specific to the ND.  It can also be run on the FD,
it's just that for the FD we're also going to always want to count tracks
in production.  However, for testing it is useful to be able to skip track
reconstruction.

Also tidy up the text in ligoanalysis.fcl.