Allow passing arbitrary arguments to nova.
In runset.sh, allow passing through commandline parameters to the
nova executable via an optional argument.  I use this to set
"-n 100", for instance, so that I don't have to process the whole
of a file.

Implementation is mildly fragile: The arguments must be quoted together
into a single script argument and must start with a "-".  This distinguishes
them from input file name, which now must not start with a "-".  But
what are you doing with files that start with "-", anwyay?  Don't do that.