Correct hit timing in MeV search.
Use speed of light in fiber to correct the timing
of hits before we search for hit pairs in the MeV
supernova-like event search.  Tighten up cuts
accordingly.  This will now run slower because it
no longer makes sense to sort hits by time, because
their time is dependent on the hit they are being
compared to.