Add code fragment for reading skymap.
Add fits.cpp, a standalone program which tests reading a LVC
skymap in FITS format, integrating the probability (carefully!)
to find the 90% CL region and printing a silly ASCII map of that
which matches the PDF the LVC provides.

This needs to be integrated into ligoanalysis_module.cc along with
something that translates (ra, dec) to a NOvA unit vector so we can
make a cut on whether a track points into the 90% CL region. The
FITS filename should be a new fcl parameter.