Get rid of analysis classes
Replace with a boolean for blindness.  The code already almost
entirely keyed off the detector of the file rather than whether
the class was SNonlyFD or SNonlyND.  The last case was for
the progress bar when reading raw files, and that can just assume
FD since it hardly matters.

Also make variable naming more consistent, remove a misleading
comment clause, and remove some unnecessary spaces.