Split cosmic time variables by view
For all of the time to/since tracks and slices, split them into
x and y view information.  The old variables are simply the
minimum of each pair.  The intent is to avoid rejecting signal
events because they are near a track or slice in one view, but
clearly separate in the other.

The variable for time since the last big shower is not split by
view because it always applies to the whole detector.

Also remove a debugging print statement.