Write out slices as clusters, add underscores
* Write out each slice in the same format as the supernova-like
clusters so that we can think about selecting events with more
than 7 hits.  Also be willing to make supernova-like clusters
with more than 7 hits for the vMM analysis.

* Make variables like x_since_footnotesize_shapeslc_s readable
by adding underscores like that.

These both break compatibility with the supernova analysis, but
only in ways that are easy to patch up: just need to cut clusters
that have slc != 0 and add underscores.